added a test for helm downloader
